Clean up data entry component
